+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.4.0] - 2020-08-30
+
 ### Fixed
 
 - Fixed misspelling of 'Throughput' in `README.md` and tests - Fixes [Issue #396](https://github.com/PlagueHO/CosmosDB/issues/396).